% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_imgpaths}
\alias{find_imgpaths}
\title{Find image paths}
\usage{
find_imgpaths(pkgnames)
}
\arguments{
\item{pkgnames}{Character vector of package names}
}
\value{
A list of image file paths for each package
}
\description{
Find image paths
}
\details{
Images in svg format will be converted to png. When no image matches 'logo'
in the file name the used is will be prompted to select likely logos.
}
