% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_logopaths}
\alias{find_logopaths}
\title{Find logo paths}
\usage{
find_logopaths(imagepaths)
}
\arguments{
\item{imagepaths}{List of image paths}
}
\value{
A vector of logo paths
}
\description{
Find logo paths
}
