% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_data.R
\name{get_pkg_data}
\alias{get_pkg_data}
\title{Get package logos and urls}
\usage{
get_pkg_data()
}
\value{
A list with paths to logos and URLs for loaded packages
}
\description{
Get package logos and urls
}
