% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_arrange.R
\name{maincolorRGB}
\alias{maincolorRGB}
\title{Extract the Most Frequent Color from an Image}
\usage{
maincolorRGB(imgpath)
}
\arguments{
\item{imgpath}{Character string. File path to the image.}
}
\value{
A data frame with one row containing the RGB values of the dominant color.
The column name is set to the input image path.
}
\description{
Internal helper. For a given image path, this functions uses k-means
clustering to identify the most dominant color in the image.
}
