% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_missingLogos}
\alias{make_missingLogos}
\title{Create missing logos}
\usage{
make_missingLogos(attached_pkgs, logopaths)
}
\arguments{
\item{attached_pkgs}{Character vector of attached package names}

\item{logopaths}{Vector of existing logo paths}
}
\value{
Vector of paths to new logos
}
\description{
Create missing logos
}
