% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pkgurls}
\alias{pkgurls}
\title{Get package URLs}
\usage{
pkgurls(pkgnames)
}
\arguments{
\item{pkgnames}{Character vector of package names}
}
\value{
A vector of package URLs
}
\description{
Get package URLs
}
