% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{encode_image}
\alias{encode_image}
\title{Encode image to Base64}
\usage{
encode_image(file_path)
}
\arguments{
\item{file_path}{Path to an image file}
}
\value{
Base64 encoded string of the image
}
\description{
Encode image to Base64
}
