% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_jsBlock.R
\name{generate_hexsession_js}
\alias{generate_hexsession_js}
\title{Generate JavaScript file for hexsession}
\usage{
generate_hexsession_js(logopaths, urls, dark_mode, output_js)
}
\arguments{
\item{logopaths}{Vector of image paths}

\item{urls}{Vector of URLs}

\item{dark_mode}{Use dark mode, inherited from make_tile}

\item{output_js}{Path to save the JavaScript file}
}
\description{
Generate JavaScript file for hexsession
}
