% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tile.R
\name{make_tile}
\alias{make_tile}
\title{Generate tile of package logos}
\usage{
make_tile(packages = NULL, local_images = NULL, dark_mode = FALSE)
}
\arguments{
\item{packages}{Character vector of package names to include (default: NULL, which uses loaded packages)}

\item{local_images}{Optional character vector of local image paths to add to the tile}

\item{dark_mode}{Draw the tile on a dark background?}

\item{local_urls}{Optional character vector of URLs for each of the local images passed}
}
\value{
Path to the output file
}
\description{
Generate tile of package logos
}
