% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_tile.R
\name{snap_tile}
\alias{snap_tile}
\title{Take screenshot of html image tile}
\usage{
snap_tile(
  output_path,
  screen_width = 800,
  screen_height = 700,
  dark_mode = FALSE
)
}
\arguments{
\item{output_path}{Path to image file}

\item{screen_width}{Width of the browser window}

\item{screen_height}{Height of the browser window}

\item{dark_mode}{Is the tile being saved dark or light mode?}
}
\value{
Path to the saved image
}
\description{
Take screenshot of html image tile
}
